# src/main/java/org/openapitools/api/ApiApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.TimeResponse;
import org.openapitools.model.UserProfileRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-03-05T22:03:09.131657600+01:00[Europe/Warsaw]")
@Validated
@Tag(name = "api", description = "the api API")
public interface ApiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * PUT /api/user/profile : Changes user profile
     *
     * @param userProfileRequest  (required)
     * @return Current time (status code 200)
     *         or Missing authentication token (status code 401)
     *         or Access forbiden (status code 403)
     */
    @Operation(
        operationId = "profile",
        summary = "Changes user profile",
        tags = { "api" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Current time", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TimeResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing authentication token"),
            @ApiResponse(responseCode = "403", description = "Access forbiden")
        },
        security = {
            @SecurityRequirement(name = "jwtToken")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/user/profile",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<TimeResponse> profile(
        @Parameter(name = "UserProfileRequest", description = "", required = true) @Valid @RequestBody UserProfileRequest userProfileRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"time\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/time : Returns current time in user time zone
     *
     * @return Current time (status code 200)
     *         or Missing authentication token (status code 401)
     *         or Access forbiden (status code 403)
     */
    @Operation(
        operationId = "time",
        summary = "Returns current time in user time zone",
        tags = { "api" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Current time", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TimeResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing authentication token"),
            @ApiResponse(responseCode = "403", description = "Access forbiden")
        },
        security = {
            @SecurityRequirement(name = "jwtToken")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/time",
        produces = { "application/json" }
    )
    default ResponseEntity<TimeResponse> time(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"time\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}